Add Select Clock option to settings app
This adds an option to select the preferred clock to show on the
main screen.

The settings app menu lists all apps of type 'clock' with the
current one marked with a *. It stores the src property
of the selected clock as settings.clock.

The bootloader app checks this property and tries to load
that app. If the property is not set, or the app not found
then it reverts to the current behaviour of using the first
clock app it finds.