Fix other app's graphics showing on image clock (BANGLEJS 1)
There is an issue, where if you open an app that has graphics on the bottom, and hold down BTN3 to go to the image clock, a small amount of the apps graphics (about 10px) will show on the bottom. This update just adds an initial clear, which clears other app's graphics.

I am using a banglejs v1 (firmware 2v21).