boot: fix `throw` stopping clock w/o wdgts loading
This surrounds the `throw` statement with a timeout and is then called
only after the clock has loaded. This seems a little weird maybe? But it
works. I also tried changing to a `conole.log` instead of the `throw`
inside a timeout.

I am not sure this is the best way to solve this. @halemmerich
@bobrippling @nxdefiant @gfwilliams do you have other suggestions? Maybe
the fix should be somewhere else completely.