Fix size of widget overlay
The widget overlay was enlarged from 24px to 26px high, to add a contrast bar at the bottom. Most of the '24' in the code were not changed to 26. In use this meant that it did not fully clear off the screen. Replaced the '24' with exports.buffer_h = 26, so they are the same size, and it clears properly.