calendar: read events synchronized from Gadgetbridge
This is a partial fix for feature request #3707.  This allows synchronized
events (stored in android.calendar.json) to be shown in the app, alongside
the manually entered events and alarm events that are already shown.  Note
only date/time and event title are currently used; other information
available in the synchronized events (e.g. duration, description) are ignored.