bwclk & bwclklite: Support 12-hour time format.
This modifies how the time string is generated so it respects the system's
locale settings, specifically for 12-hour time.  This is done with the
locale package, which also simplifies some of the string handling since the
package creates a complete time string.  Note this does not add any display of
AM/PM, but neither do any other clock faces.