messagegui: avoid use of the overview screen
make many tweaks to accommodate that change.

Add external back on touch handler to the showMessagesScroller that
shows the list of messages.

Select a message (touch it on B2, HW button on B1) opens the message
settings screen.

Going back from message setting screen opens the showMessagesScroller on
the current message again.

TODO: See if the code for the showMessageOverview can be removed.